lotus/client: refactor
